requirement.txt
# Packages listed in alphabetical order
numpy
bokeh
opencv-python
pytest
pyyaml
scikit-learn
streamlit
trimesh
pyrender
open3d
pywavefront
scipy
mediapipe
opencv-python
torch
diffusers
transformers
accelerate
safetensors
numpy
huggingface
Pillow
PyOpenGL 
PyOpenGL_accelerate
open3d
pygame
# pytorch special case, see requirements-pytorch.txt